+++ main.cpp
@@ -14,13 +14,13 @@ int main() {
     cout << "Digite a palavra a ser buscada: ";
     getline(cin, palavra);
 
-    vector<int> posicoes = buscarKMP(texto, palavra);
+    vector<size_t> posicoes = buscarKMP(texto, palavra);
 
     if (posicoes.empty()) {
         cout << "Palavra não encontrada no texto." << endl;
     } else {
         cout << "Palavra encontrada nas posições: ";
-        for (int pos : posicoes) {
+        for (size_t pos : posicoes) {
             cout << pos << " ";
         }
         cout << endl;
